+++ mutalyzer/__init__.py
@@ -21,8 +21,8 @@ from __future__ import unicode_literals
 # [1] http://peak.telecommunity.com/DevCenter/setuptools#specifying-your-project-s-version
 # [2] http://semver.org/
 
-__version_info__ = ('2', '0', '7', 'dev')
-__date__ = '10 Feb 2015'
+__version_info__ = ('2', '0', '8', 'dev')
+__date__ = '12 Mar 2015'
 
 
 __version__ = '.'.join(__version_info__)